Add files via upload
Makeup Application is a virtual market place on the internet where customer can browse the list and select product of interest. The selected item will be collected in a shopping cart, at checkout time it will be presented as an order. The customer will be ask to fill address and payment information. As soon as the order is placed a pop up message displays to the customer.     
( https://mymakeup.pythonanywhere.com/ )